Fix the build of `readelf` on 32-bit systems.
Reviewed by:	kaiwang27

git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@2740 95820547-d848-0410-985e-9ae8fe0fa350

+++ readelf/readelf.c
@@ -3470,7 +3470,8 @@ dump_verdef(struct readelf *re, int dump)
 				SAVE_VERSION_NAME((int)vd->vd_ndx, name, 1);
 			} else if (dump)
 				printf("  0x%4.4lx parent: %s\n",
-				    buf2 - (uint8_t *)d->d_buf, name);
+				    (unsigned long) (buf2 -
+				    (uint8_t *)d->d_buf), name);
 			if (vda->vda_next == 0)
 				break;
 			buf2 += vda->vda_next;
